refactor(package): rename posts package to articles package
Rename posts package to article packages to avoid names collision and misunderstandings
