+++ app/src/main/java/com/makentoshe/habrachan/model/main/articles/ArticlesFlowScreen.kt
@@ -1,8 +1,8 @@
-package com.makentoshe.habrachan.model.main.posts
+package com.makentoshe.habrachan.model.main.articles
 
 import androidx.fragment.app.Fragment
 import com.makentoshe.habrachan.common.navigation.Screen
-import com.makentoshe.habrachan.view.main.posts.ArticlesFlowFragment
+import com.makentoshe.habrachan.view.main.articles.ArticlesFlowFragment
 
 class ArticlesFlowScreen(private val page: Int) : Screen(){
     override val fragment: Fragment
