+++ app/src/main/java/com/makentoshe/habrachan/view/main/MainFlowFragment.kt
@@ -12,8 +12,8 @@ import com.makentoshe.habrachan.di.common.ApplicationScope
 import com.makentoshe.habrachan.di.main.MainFlowFragmentModule
 import com.makentoshe.habrachan.di.main.MainFlowFragmentScope
 import com.makentoshe.habrachan.model.main.account.AccountFlowScreen
+import com.makentoshe.habrachan.model.main.articles.ArticlesFlowScreen
 import com.makentoshe.habrachan.model.main.menu.MenuScreen
-import com.makentoshe.habrachan.model.main.posts.ArticlesFlowScreen
 import com.makentoshe.habrachan.ui.main.MainFlowFragmentUi
 import ru.terrakok.cicerone.NavigatorHolder
 import ru.terrakok.cicerone.Router
