+++ app/src/main/res/layout/articles_flow_fragment.xml
@@ -49,7 +49,7 @@
 
         <fragment
             android:id="@+id/articles_fragment"
-            android:name="com.makentoshe.habrachan.view.main.posts.ArticlesFragment"
+            android:name="com.makentoshe.habrachan.view.main.articles.ArticlesFragment"
             android:layout_width="match_parent"
             android:layout_height="match_parent" />
